requirements.txt
pandas
jupyter
jupyter-client
seaborn
numpy
matplotlib
nltk
scikit-learn
tensorflow-gpu==2.7.1
imageai==2.1.6
opencv-python==4.5.5.62
requests
pillow==9.1.0
tqdm
deepface
google-api-python-client
imbalanced-learn
featurewiz